requirements.txt
dropbox==11.36.2
google_api_python_client==2.77.0
matplotlib==3.6.2
netCDF4==1.7.2
numpy==1.24.1
Pillow==10.1.0
protobuf==5.29.2
python-decouple==3.8
Requests==2.32.3
